pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "BigTime ChatBT AI assistant"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
pandas = "^2.1.2"
openai = "^1.14.0"
pinecone-client = "^3.1.0"
gunicorn = "^21.2.0"
PyMongo = "3.12.3"
flask-session = "^0.6.0"
tiktoken = "^0.5.1"
flask = {extras = ["async"], version = "^3.0.3"}
werkzeug = "3.0.1"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
opentelemetry-api = "^1.23.0"
opentelemetry-sdk = "^1.23.0"
opentelemetry-instrumentation-flask = "^0.45b0"
opentelemetry-instrumentation-requests = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
anthropic = "^0.21.3"
groq = "^0.5.0"
opentelemetry-distro = "^0.45b0"
flask-cors = "^4.0.1"
duckdb = "^1.0.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
toml = "^0.10.2"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"